requirements.txt
crewai>=1.3.0
crewai-tools>=1.3.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
openai>=1.35.0
python-dotenv>=1.0.1
chromadb-client>=0.5.3